refactor(types): retire duplicate type authority
