+++ test/integration/bible-lookup.test.ts
@@ -7,7 +7,7 @@
 
 import { describe, it, expect, beforeEach } from 'vitest';
 import { bibleLookupHandler } from '../../src/tools/bibleLookup.js';
-import type { BibleLookupParams } from '../../src/types/index.js';
+import type { BibleLookupParams } from '../../src/kernel/types.js';
 
 describe('Bible Lookup Integration Tests', () => {
   describe('Single Translation Lookups', () => {
