+++ test/scripts/bible-lookup-17-tests.ts
@@ -11,7 +11,7 @@
  */
 
 import { bibleLookupHandler } from '../../src/tools/bibleLookup.js';
-import type { BibleLookupParams } from '../../src/types/index.js';
+import type { BibleLookupParams } from '../../src/kernel/types.js';
 
 interface TestResult {
   testNumber: number;
